# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "API for an employee management system to determine employees that\
    \ are no longer working on active projects and to understand their qualifications\
    \ and expertise for suitability in other projects"
  version: "1.0.0"
  title: "UBahn API"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /users/{userId}/skills:
    get:
      tags:
      - "Users Skills"
      description: "Filter skills by its name given an user id.  If no results, then\
        \ empty array is returned.\n\n**Security** - Note that for non-admin users,\
        \ this endpoint will only return entities that\nthe user has created.\n"
      operationId: "usersUserIdSkillsGET"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "skillName"
        in: "query"
        description: "Filter by skill name (through skill id)"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UserSkill"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "UsersSkills"
    head:
      tags:
      - "Users Skills"
      description: "Retrieve header information for a search operation on users skills\
        \ in the application.\n\n**Security** - Note that for non-admin users, this\
        \ endpoint will only return entities that\nthe user has created.\n"
      operationId: "usersUserIdSkillsHEAD"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "skillName"
        in: "query"
        description: "Filter by skill name (through skill id)"
        required: false
        type: "string"
      responses:
        "200":
          description: "Success response"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "UsersSkills"
    post:
      tags:
      - "Users Skills"
      description: "Create a new User Skill.\n\n**Security** - This endpoint is accessible\
        \ by all authenticated users.        \n"
      operationId: "usersUserIdSkillsPOST"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/UserSkillRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/UserSkill"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "UsersSkills"
  /users/{userId}/skills/{skillId}:
    get:
      tags:
      - "Users Skills"
      description: "Get User Skills with given user and skill id.\n\n**Security**\
        \ - Note that for non-admin users, this endpoint will only return entities\
        \ that\nthe user has created.        \n"
      operationId: "usersUserIdSkillsSkillIdGET"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "skillId"
        in: "path"
        description: "The skill id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/UserSkill"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "UsersSkills"
    head:
      tags:
      - "Users Skills"
      description: "Get User Skills with given ids, but only header information is\
        \ returned.\n\n**Security** - Note that for non-admin users, this endpoint\
        \ will only return entities that\nthe user has created.        \n"
      operationId: "usersUserIdSkillsSkillIdHEAD"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "skillId"
        in: "path"
        description: "The skill id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "UsersSkills"
    delete:
      tags:
      - "Users Skills"
      description: "Remove an existing User Skill with given ids.\n\n**Security**\
        \ - Note that this endpoint is only available for admin users.  \n"
      operationId: "usersUserIdSkillsSkillIdDELETE"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "skillId"
        in: "path"
        description: "The skill id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "204":
          description: "OK - the request was successful"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "UsersSkills"
    patch:
      tags:
      - "Users Skills"
      description: "Update an existing skill with given ids.\n\n**Security** - Note\
        \ that for non-admin users, this endpoint will only allow updates on entities\
        \ that the\ncalling user has created.\n"
      operationId: "usersUserIdSkillsSkillIdPATCH"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "skillId"
        in: "path"
        description: "The skill id"
        required: true
        type: "string"
        format: "UUID"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/UserSkillUpdateRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/UserSkill"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "UsersSkills"
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  AuditFields:
    type: "object"
    required:
    - "created"
    - "createdBy"
    - "updated"
    - "updatedBy"
    properties:
      created:
        type: "string"
        format: "date-time"
        description: "When the entity was created."
      updated:
        type: "string"
        format: "date-time"
        description: "When the entity was updated."
      createdBy:
        type: "string"
        format: "UUID"
        description: "Creator of the entity."
      updatedBy:
        type: "string"
        format: "UUID"
        description: "User that last updated the entity."
    description: "Describes the audit fields that are present in all the models in\
      \ this API."
...
  UserSkill:
    allOf:
    - type: "object"
      required:
      - "certifiedDate"
      - "certifierId"
      - "metricValue"
      - "skillId"
      - "userId"
      properties:
        userId:
          type: "string"
          format: "UUID"
          description: "The id of user that this Skill belongs to."
        skillId:
          type: "string"
          format: "UUID"
          description: "The Skill id."
        metricValue:
          type: "string"
          description: "The skill metric value."
        certifierId:
          type: "string"
          description: "Id of certifier"
        certifiedDate:
          type: "string"
          format: "date-time"
          description: "The date when certification occurred."
    - $ref: "#/definitions/AuditFields"
  UserSkillRequestBody:
    allOf:
    - type: "object"
      properties:
        skillId:
          type: "string"
          format: "UUID"
          description: "The Skill id."
    - $ref: "#/definitions/UserSkillUpdateRequestBody"
  UserSkillUpdateRequestBody:
    type: "object"
    properties:
      metricValue:
        type: "string"
        description: "The skill metric value."
      certifierId:
        type: "string"
        description: "Id of certifier"
      certifiedDate:
        type: "string"
        format: "date-time"
        description: "The date when certification occurred."
    example:
      certifierId: "certifierId"
      certifiedDate: "2000-01-23T04:56:07.000+00:00"
      metricValue: "metricValue"
...
  Unauthorized:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Unable to authenticate the user."
        description: "The unauthorized error message."
    description: "The unauthorized error entity."
  NotFound:
    type: "object"
    properties:
      message:
        type: "string"
        example: "A resource with the name could not be found."
        description: "The not found error message."
    description: "The not found error entity."
  ServerError:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Something went wrong while processing your request. We�re sorry\
          \ for the trouble. We�ve been notified of the error and will correct it\
          \ as soon as possible. Please try your request again in a moment."
        description: "The server error message."
    description: "The server error entity."
  BadRequest:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Invalid input."
        description: "The bad request error message."
    description: "The bad request error entity."
  Forbidden:
    type: "object"
    properties:
      message:
        type: "string"
        example: "You are not allowed to access the request."
        description: "The forbidden error message."
    description: "The permission error entity."
  Conflict:
    type: "object"
    required:
    - "message"
    properties:
      message:
        type: "string"
        example: "Creating a resource with a name already exists."
        description: "The conflict error message."
    description: "The conflict error entity."